added setting env variable to the current branch you are on.
added command execution that gets and sets the current branch you are working on. so that when you access the yml files you access them in the correct branch automatically.

changed the proccess variable that defines the "IP address" in which the server is running on,
in development that is ngrok,
production it will be the domain that the server is hostes on.
changed it to be IP_ADDRESS instead and the previous refferences to "MY_URL" reffer to IP_ADDRESS instead